split up into two tasks for heroku
