+++ .gitignore
@@ -13,6 +13,7 @@ capybara-*.html
 **.orig
 rerun.txt
 routes.txt
+people.csv
 pickle-email-*.html
 
 # dotenv
