+++ app/use_cases/echo_people.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+require 'csv' # Not a gem: in ruby stdlib
+
+class EchoPeople < ActiveInteraction::Base
+  string :file_name
+
+  # Loads people from the given file name
+  #
+  # Params:
+  # - file_name: the name of the csv file to load
+  def execute
+    assert_file_existence(file_name)
+    echo_people(file_name)
+  end
+
+  private
+
+  def assert_file_existence(file_name)
+    raise "File #{file_name} does not exist" unless File.file?(file_name)
+  end
+
+  def echo_people(file_name)
+    people = []
+    cnt = 0
+    puts 'people = [];nil'
+    CSV.foreach(file_name, col_sep: ';') do |row|
+      cnt += 1
+      next if cnt == 1 # Skip header row
+      people << { first_name: row[0],
+                  last_name: row[1],
+                  mobile_phone: row[2],
+                  protocol_name: row[3],
+                  start_date: row[4] }
+      puts "people << #{people.last.inspect};nil"
+    end
+  end
+end
