+++ lib/tasks/maintenance.rake
@@ -2,11 +2,11 @@
 
 namespace :maintenance do
   # Run with
-  # be rake "maintenance:load_people[people.csv]"
-  desc 'Load people from a csv file'
-  task :load_people, [:csv_file] => [:environment] do |_, args|
-    puts "Loading people from '#{args[:csv_file]}' - started"
-    LoadPeople.run!(file_name: args[:csv_file])
-    puts "Loading people from '#{args[:csv_file]}' - done"
+  # be rake "maintenance:echo_people[people.csv]"
+  desc 'Echo people from a csv file'
+  task :echo_people, [:csv_file] => [:environment] do |_, args|
+    puts "Echoing people from '#{args[:csv_file]}' - started"
+    EchoPeople.run!(file_name: args[:csv_file])
+    puts "Echoing people from '#{args[:csv_file]}' - done"
   end
 end
